COMP: Drop redundant int cast in vnl_rnpoly_solve powr() calls
powr() takes an unsigned int and ideg is std::vector<unsigned int>, so
casting ideg[j] down to int before the call is unnecessary and forces a
sign-conversion round-trip (unsigned -> int -> unsigned). Pass ideg[j]
directly.

+++ core/vnl/algo/vnl_rnpoly_solve.cxx
@@ -195,8 +195,8 @@ initr(const std::vector<unsigned int> & ideg,
   r.resize(dim_);
   for (unsigned int j = 0; j < dim_; j++)
   {
-    pdg[j] = powr(static_cast<int>(ideg[j]), p[j]);
-    qdg[j] = powr(static_cast<int>(ideg[j]), q[j]);
+    pdg[j] = powr(ideg[j], p[j]);
+    qdg[j] = powr(ideg[j], q[j]);
     r[j] = q[j] / p[j];
   }
 }
